requirements.txt
PyJWT==1.7.1
apache-airflow~=1.10.10
arrow~=0.15.4
django~=2.2.3
django-crontab~=0.7.1
django-filter~=2.2.0
django-localflavor-us~=1.1
djangorestframework~=3.9.4
django-cors-headers~=3.0.2
django-graphene-permissions~=0.0.4
django-graphql-social-auth[jwt]~=0.1.4
graphene-django~=2.10.1
graphene-graphiql-explorer~=0.0.1
gunicorn~=19.9.0
markdown~=2.6.10
openpyxl~=3.0.5
pandas~=1.2.2
psycopg2~=2.8.2
psycopg2-binary~=2.8.3
sendgrid~=6.2.1
stripe>=2.48.0,<3.0
twilio~=6.41.0
django-graphql-jwt~=0.3.1
graphene_file_upload~=1.3.0
openpyxl~=1.0.1